requirements.txt
SQLAlchemy~=1.3.13
pycryptodome~=3.9.0
websockets~=9.1
flask~=1.1.2
Flask-WTF~=0.15.1
Flask-Bootstrap4==4.0.2
python-socketio==5.3.0
eventlet~=0.31.0
xlsxwriter~=1.4.4
psutil~=5.8.0; platform_system=="Windows"
loguru
python-dotenv
